Add startup WoR query
