+++ lib/views/setup.js
@@ -1,18 +1,12 @@
 
 var pug = require('pug')
-
 var extend = require('xtend')
-
 var config = require('../config')
-
+var getRegistries = require('../wor')
 var validator = require('validator');
-
 var createUser = require('../createUser')
-
 const uuid = require('uuid/v4')
-
 const theme = require('../theme')
-
 const fs = require('fs')
 
 module.exports = function (req, res) {
@@ -144,7 +138,7 @@ function setupPost(req, res, settings) {
     config.set('shareLinkSalt', uuid())
     config.set('passwordSalt', uuid())
     config.set('instanceUrl', settings.instanceUrl)
-    config.set('webOfRegistries', {[settings.instanceUrl]: settings.instanceUrl})
+    config.set('webOfRegistries', {[settings.uriPrefix.slice(0, -1)]: settings.instanceUrl.slice(0, -1)})
     config.set('databasePrefix', settings.instanceUrl)
     config.set('themeParameters', { 'default': { baseColor: settings.color } })
     config.set('frontPageText', settings.frontPageText)
@@ -178,6 +172,8 @@ function setupPost(req, res, settings) {
 
         return theme.setCurrentThemeFromConfig()
 
+    }).then(() => {
+        return getRegistries()
     }).then(() => {
         req.session.save(() => {
             res.redirect(req.body.next || '/');
