+++ package.json
@@ -6,6 +6,7 @@
     "alouette": "^5.5.0",
     "async": "^0.9.2",
     "autocomplete": "0.0.1",
+    "axios": "^0.18.0",
     "basic-auth": "^1.0.3",
     "binary-split": "^1.0.3",
     "body-parser": "^1.15.0",
